AsyncLocalStorage based ContextManager (#1210)
